Attempt to set up GOOL tests (not working)
