+++ code/CodeGen/codegen.cabal
@@ -1,0 +1,33 @@
+Name:		codegen
+Version:	0.0.1
+Cabal-Version:  >= 1.18
+Author:		Brooks MacLachlan
+build-type:     Simple
+Synopsis:	Tests of finally tagless GOOL
+
+executable codegenTest
+  hs-source-dirs: .
+  main-is: Test/Main.hs
+  other-modules:
+      Test.HelloWorld
+    , Test.FileTests
+    , Test.PatternTest
+    , Test.Observer
+    , Test.Helper
+  Build-Depends:
+    base >= 4.7,
+    lens >= 4.12,
+    containers >= 0.5.5.1,
+    pretty >= 1.1.1.1,
+    mtl >= 2.2.1,
+    filepath >= 1.4.1,
+    directory >= 1.2.6.2,
+    split >= 0.2.3.1,
+    drasil-code >= 0.1.4
+  default-language: Haskell2010
+  ghc-options:      -Wall -O2 -Wredundant-constraints
+
+source-repository head
+  type:     git
+  location: git://github.com/JacquesCarette/Drasil.git
+  subdir: code
