+++ code/CodeGen/stack.yaml
@@ -12,13 +12,10 @@
 # resolver: nightly-2015-09-21
 # resolver: ghc-7.10.2
 # resolver: ghcjs-0.1.0_ghc-7.10.2
-#
-# The location of a snapshot can be provided as a file or url. Stack assumes
-# a snapshot provided as a file might change, whereas a url resource does not.
-#
-# resolver: ./custom-snapshot.yaml
-# resolver: https://example.com/snapshots/2018-01-01.yaml
-resolver: lts-12.22
+# resolver:
+#  name: custom-snapshot
+#  location: "./custom-snapshot.yaml"
+resolver: lts-11.2
 
 # User packages to be built.
 # Various formats can be used as shown in the example below.
@@ -30,13 +27,22 @@ resolver: lts-12.22
 #    git: https://github.com/commercialhaskell/stack.git
 #    commit: e7b331f14bcffb8367cd58fbfc8b40ec7642100a
 # - location: https://github.com/commercialhaskell/stack/commit/e7b331f14bcffb8367cd58fbfc8b40ec7642100a
+#   extra-dep: true
 #  subdirs:
 #  - auto-update
 #  - wai
+#
+# A package marked 'extra-dep: true' will only be built if demanded by a
+# non-dependency (i.e. a user package), and its test suites and benchmarks
+# will not be run. This is useful for tweaking upstream packages.
 packages:
-- .
+- ../drasil-lang
+- ../drasil-build
+- ../drasil-database
+- ../drasil-code
+- ../drasil-theory
+- ./
 # Dependency packages to be pulled from upstream that are not in the resolver
-# using the same syntax as the packages field.
 # (e.g., acme-missiles-0.3)
 # extra-deps: []
 
@@ -51,7 +57,7 @@ packages:
 #
 # Require a specific version of stack, using version ranges
 # require-stack-version: -any # Default
-# require-stack-version: ">=1.7"
+# require-stack-version: ">=1.6"
 #
 # Override the architecture used by stack, especially useful on Windows
 # arch: i386
