+++ code/drasil-code/drasil-code.cabal
@@ -32,7 +32,7 @@ library
     pretty >= 1.1.1.1,
     mtl >= 2.2.1,
     directory >= 1.2.6.2,
-    filepath >= 1.4.2,
+    filepath >= 1.4.1,
     split >= 0.2.3.1,
     MissingH >= 1.4.0.1,
     parsec >= 3.1.9,
